@@ -3,16 +3,13 @@ package com.hmdp.config;
 import com.hmdp.utils.LoginInterceptor;
 import com.hmdp.utils.RefreshTokenInteceptor;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.data.redis.core.StringRedisTemplate;
-import org.springframework.stereotype.Controller;
 import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 
-import javax.annotation.Resource;
+import jakarta.annotation.Resource;
 
 @Configuration
 public class MvcConfig implements WebMvcConfigurer {
-
 
     @Resource
     private LoginInterceptor loginInterceptor;
@@ -24,10 +21,9 @@ public class MvcConfig implements WebMvcConfigurer {
         registry.addInterceptor(loginInterceptor)
                 .excludePathPatterns(
                         "/user/login",
-                        "/user/code"
-
-
-                ).order(1);
+                        "/user/code",
+                        "/chat")
+                .order(1);
         registry.addInterceptor(refreshTokenInteceptor).order(0);
     }
 }
