@@ -11,8 +11,8 @@ import com.hmdp.utils.UserHolder;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.web.bind.annotation.*;
 
-import javax.annotation.Resource;
-import javax.servlet.http.HttpSession;
+import jakarta.annotation.Resource;
+import jakarta.servlet.http.HttpSession;
 
 
 @Slf4j
@@ -37,7 +37,7 @@ public class UserController {
 
     /**
      * 登录功能
-     * @param loginForm 登录参数，包含手机号、验证码；或者手机号、密码
+     * @param loginForm 登录参数，包含手机号、验证码；或者手机号、密�?
      */
     @PostMapping("/login")
     public Result login(@RequestBody LoginFormDTO loginForm, HttpSession session){
@@ -47,12 +47,12 @@ public class UserController {
 
     /**
      * 登出功能
-     * @return 无
+     * @return �?
      */
     @PostMapping("/logout")
     public Result logout(){
         // TODO 实现登出功能
-        return Result.fail("功能未完成");
+        return Result.fail("功能未完善");
     }
 
     @GetMapping("/me")
@@ -67,7 +67,7 @@ public class UserController {
         // 查询详情
         UserInfo info = userInfoService.getById(userId);
         if (info == null) {
-            // 没有详情，应该是第一次查看详情
+            // 没有详情，应该是第一次查看详�?
             return Result.ok();
         }
         info.setCreateTime(null);
