@@ -6,7 +6,7 @@ import com.hmdp.entity.Voucher;
 import com.hmdp.service.IVoucherService;
 import org.springframework.web.bind.annotation.*;
 
-import javax.annotation.Resource;
+import jakarta.annotation.Resource;
 
 
 @RestController
@@ -18,7 +18,7 @@ public class VoucherController {
 
     /**
      * 新增普通券
-     * @param voucher 优惠券信息
+     * @param voucher 优惠券信�?
      * @return 优惠券id
      */
     @PostMapping
@@ -28,7 +28,7 @@ public class VoucherController {
     }
 
     /**
-     * 新增秒杀券
+     * 新增秒杀�?
      * @param voucher 优惠券信息，包含秒杀信息
      * @return 优惠券id
      */
@@ -41,7 +41,7 @@ public class VoucherController {
     /**
      * 查询店铺的优惠券列表
      * @param shopId 店铺id
-     * @return 优惠券列表
+     * @return 优惠券列�?
      */
     @GetMapping("/list/{shopId}")
     public Result queryVoucherOfShop(@PathVariable("shopId") Long shopId) {
