@@ -5,11 +5,11 @@ import com.hmdp.dto.LoginFormDTO;
 import com.hmdp.dto.Result;
 import com.hmdp.entity.User;
 
-import javax.servlet.http.HttpSession;
+import jakarta.servlet.http.HttpSession;
 
 /**
  * <p>
- *  服务类
+ *  服务�?
  * </p>
  *
  * @author 虎哥
@@ -27,7 +27,7 @@ public interface IUserService extends IService<User> {
 
     /**
      * 登录功能
-     * @param loginForm 登录参数，包含手机号、验证码；或者手机号、密码
+     * @param loginForm 登录参数，包含手机号、验证码；或者手机号、密�?
      */
     Result login(LoginFormDTO loginForm, HttpSession session);
 }
