@@ -5,7 +5,7 @@ import com.baomidou.mybatisplus.extension.service.IService;
 
 /**
  * <p>
- * 秒杀优惠券表，与优惠券是一对一关系 服务类
+ * 秒杀优惠券表，与优惠券是一对一关系 服务�?
  * </p>
  *
  * @author 虎哥
