@@ -5,15 +5,15 @@ import cn.hutool.core.util.StrUtil;
 
 public class RegexUtils {
     /**
-     * 是否是无效手机格式
-     * @param phone 要校验的手机号
+     * 是否是无效手机格�?
+     * @param phone 要校验的手机�?
      * @return true:符合，false：不符合
      */
     public static boolean isPhoneInvalid(String phone){
         return mismatch(phone, RegexPatterns.PHONE_REGEX);
     }
     /**
-     * 是否是无效邮箱格式
+     * 是否是无效邮箱格�?
      * @param email 要校验的邮箱
      * @return true:符合，false：不符合
      */
@@ -23,14 +23,14 @@ public class RegexUtils {
 
     /**
      * 是否是无效验证码格式
-     * @param code 要校验的验证码
+     * @param code 要校验的验证�?
      * @return true:符合，false：不符合
      */
     public static boolean isCodeInvalid(String code){
         return mismatch(code, RegexPatterns.VERIFY_CODE_REGEX);
     }
 
-    // 校验是否不符合正则格式
+    // 校验是否不符合正则格�?
     private static boolean mismatch(String str, String regex){
         if (StrUtil.isBlank(str)) {
             return true;
